src/Backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
requests

